// packages/cc/src/lib/_Types.d.ts
export declare enum EntryControlCommand {
    Notification = 1,
    KeySupportedGet = 2,
    KeySupportedReport = 3,
    EventSupportedGet = 4,
    EventSupportedReport = 5,
    ConfigurationSet = 6,
    ConfigurationGet = 7,
    ConfigurationReport = 8
}
export declare enum EntryControlDataTypes {
    None = 0,
    Raw = 1,
    ASCII = 2,
    MD5 = 3
}